Allow subscription to ngRedux in @Injectable
When trying to use `ngRedux.select()` in constructor code in your
services, there was a chance that the store had not been configured yet
and would throw errors.

Updated code so store$ starts off as an initalized BehaviorSubject, and
the exposed streams off of it will filter for 'null' values first,
ensuring that downstream subscribers will have a proper observable, and
also the correct version of app state.